libs: st7789: Use DMA to improve rect drawing.

+++ STM32-CMSIS-Libs/libs/st7789/driver_st7789_interface.h
@@ -38,6 +38,7 @@
 #define DRIVER_ST7789_INTERFACE_H
 
 #include "Driver_SPI.h"
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C"{
@@ -77,7 +78,7 @@ uint8_t st7789_interface_spi_deinit(void);
  *            - 1 write failed
  * @note      none
  */
-uint8_t st7789_interface_spi_write_cmd(uint8_t *buf, uint16_t len);
+uint8_t st7789_interface_spi_write_cmd(uint8_t *buf, uint32_t len);
 
 /**
  * @brief     interface delay ms
